Move ajax-request tests into mixin folder
Also, expect-ify the test assertions
